refactor(report): move utility modules to an utility folder
