+++ src/report/utl/dependencyToIncidenceTransformer.js
